// src/interfaces/xer-types.ts
export type Task = {
  task_id: string
  proj_id: string
  wbs_id: string
  clndr_id: string
  phys_complete_pct: string
  rev_fdbk_flag: string
  est_wt: string
  lock_plan_flag: string
  auto_compute_act_flag: string
  complete_pct_type: string
  task_type: string
  duration_type: string
  status_code: string
  task_code: string
  task_name: string
  rsrc_id: string
  total_float_hr_cnt: string
  free_float_hr_cnt: string
  remain_drtn_hr_cnt: string
  act_work_qty: string
  remain_work_qty: string
  target_work_qty: string
  target_drtn_hr_cnt: string
  target_equip_qty: string
  act_equip_qty: string
  remain_equip_qty: string
  cstr_date: string
  act_start_date: string
  act_end_date: string
  late_start_date: string
  late_end_date: string
  expect_end_date: string
  early_start_date: string
  early_end_date: string
  restart_date: string
  reend_date: string
  target_start_date: string
  target_end_date: string
  rem_late_start_date: string
  rem_late_end_date: string
  cstr_type: string
  priority_type: string
  suspend_date: string
  resume_date: string
  float_path: string
  float_path_order: string
  guid: string
  tmpl_guid: string
  cstr_date2: string
  cstr_type2: string
  driving_path_flag: string
  act_this_per_work_qty: string
  act_this_per_equip_qty: string
  external_early_start_date: string
  external_late_end_date: string
  create_date: string
  update_date: string
  create_user: string
  update_user: string
  location_id: string
}

export type MappedTask = {
  start: Date
  end: Date
  targetStart: Date
  targetEnd: Date
  name: string
}